Merget branches og oppdatert struktur
