+++ oppgavetekster/Oppgave4.md
